chore: move ollama from `llms/ollama` to `ollama`
