+++ desktop_app/src/backend/ollama/client.ts
