+++ desktop_app/src/backend/ollama/index.ts
