+++ desktop_app/src/backend/ollama/server.test.ts
