+++ desktop_app/src/backend/server/plugins/ollama/metadata.ts
@@ -2,7 +2,7 @@ import type { FastifyPluginAsyncZod } from 'fastify-type-provider-zod';
 import { z } from 'zod';
 
 import config from '@backend/config';
-import OllamaClient from '@backend/llms/ollama/client';
+import OllamaClient from '@backend/ollama/client';
 
 const {
   ollama: { requiredModels: ollamaRequiredModels },
